requirements.txt
greenlet==2.0.2
pandas==2.0.1
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.12

